requirements.txt
flask
requests
pandas
numpy
jupyter
python-dotenv
